requirements.txt
pillow
opencv-python
pytesseract
pynput
pyautogui
numpy
pywin32
pandas
google-genai